Create JsonFormBuilder class.
